# docs/apispec/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API仕様書",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "ローカル環境"
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "tags": [
          "Book"
        ],
        "operationId": "searchBooks",
        "parameters": [
          {
            "name": "isbn",
            "in": "query",
            "description": "ISBN-13:完全一致",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": 9784873115658
          },
          {
            "name": "title",
            "in": "query",
            "description": "書籍タイトル:部分一致",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priceFrom",
            "in": "query",
            "description": "価格:下限",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "priceTo",
            "in": "query",
            "description": "価格:上限",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "ステータス",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "UNPUBLISHED",
                  "PUBLISHED",
                  "OUT_OF_PRINT"
                ]
              }
            }
          },
          {
            "name": "publishedAtStart",
            "in": "query",
            "description": "出版日時:ISO8601:開始",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2025-01-23T01:23:45.000Z"
          },
          {
            "name": "publishedAtEnd",
            "in": "query",
            "description": "出版日時:ISO8601:終了",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2025-01-23T01:23:45.000Z"
          },
          {
            "name": "order",
            "in": "query",
            "description": "並び順",
            "required": false,
            "schema": {
              "type": "string",
              "default": "-published_at",
              "enum": [
                "+price",
                "-price",
                "+published_at",
                "-published_at"
              ]
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "取得開始位置",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "取得数",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchBooksResponse"
                }
              }
            }
          },
          "400": {
            "description": "不正なリクエスト",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "アプリケーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Book"
        ],
        "operationId": "createBook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "成功"
          },
          "400": {
            "description": "不正なリクエスト",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "アプリケーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "tags": [
          "Book"
        ],
        "operationId": "findBookById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindBookByIdResponse"
                }
              }
            }
          },
          "400": {
            "description": "不正なリクエスト",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "存在しないリソース",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "アプリケーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Book"
        ],
        "operationId": "deleteBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "成功"
          },
          "400": {
            "description": "不正なリクエスト",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "存在しないリソース",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "アプリケーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Book"
        ],
        "operationId": "updateBook",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBookBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功"
          },
          "400": {
            "description": "不正なリクエスト",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "存在しないリソース",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "アプリケーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/books/download": {
      "get": {
        "tags": [
          "Book"
        ],
        "operationId": "downloadBooks",
        "parameters": [
          {
            "name": "isbn",
            "in": "query",
            "description": "ISBN-13:完全一致",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": 9784873115658
          },
          {
            "name": "title",
            "in": "query",
            "description": "書籍タイトル:部分一致",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priceFrom",
            "in": "query",
            "description": "価格:下限",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "priceTo",
            "in": "query",
            "description": "価格:上限",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "ステータス",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "UNPUBLISHED",
                  "PUBLISHED",
                  "OUT_OF_PRINT"
                ]
              }
            }
          },
          {
            "name": "publishedAtStart",
            "in": "query",
            "description": "出版日時:ISO8601:開始",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2025-01-23T01:23:45.000Z"
          },
          {
            "name": "publishedAtEnd",
            "in": "query",
            "description": "出版日時:ISO8601:終了",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2025-01-23T01:23:45.000Z"
          },
          {
            "name": "order",
            "in": "query",
            "description": "並び順",
            "required": false,
            "schema": {
              "type": "string",
              "default": "-published_at",
              "enum": [
                "+price",
                "-price",
                "+published_at",
                "-published_at"
              ]
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "取得開始位置",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "取得数",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          },
          "400": {
            "description": "不正なリクエスト",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "アプリケーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateBookBody": {
        "type": "object",
        "properties": {
          "isbn": {
            "type": "string",
            "description": "ISBN-13",
            "example": 9784873115658,
            "minLength": 1
          },
          "title": {
            "type": "string",
            "description": "書籍タイトル",
            "example": "リーダブルコード",
            "minLength": 1
          },
          "price": {
            "type": "integer",
            "format": "int32",
            "description": "価格",
            "example": 2640,
            "minimum": 0
          },
          "status": {
            "type": "string",
            "description": "ステータス",
            "enum": [
              "UNPUBLISHED",
              "PUBLISHED",
              "OUT_OF_PRINT"
            ],
            "example": "PUBLISHED"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "description": "出版日時:ISO8601",
            "example": "2025-01-23T01:23:45.000Z"
          }
        },
        "required": [
          "isbn",
          "status",
          "title"
        ]
      },
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {}
          }
        }
      },
      "UpdateBookBody": {
        "type": "object",
        "properties": {
          "isbn": {
            "type": "string",
            "description": "ISBN-13",
            "example": 9784873115658
          },
          "title": {
            "type": "string",
            "description": "書籍タイトル",
            "example": "リーダブルコード"
          },
          "price": {
            "type": "integer",
            "format": "int32",
            "description": "価格",
            "example": 2640,
            "minimum": 0
          },
          "status": {
            "type": "string",
            "description": "ステータス",
            "enum": [
              "UNPUBLISHED",
              "PUBLISHED",
              "OUT_OF_PRINT"
            ],
            "example": "PUBLISHED"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "description": "出版日時:ISO8601",
            "example": "2025-01-23T01:23:45.000Z"
          }
        }
      },
      "ResponseHit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "書籍ID",
            "example": 1
          },
          "isbn": {
            "type": "string",
            "description": "ISBN-13",
            "example": 9784873115658
          },
          "title": {
            "type": "string",
            "description": "タイトル",
            "example": "リーダブルコード"
          },
          "price": {
            "type": "integer",
            "format": "int32",
            "description": "価格",
            "example": 2640
          },
          "status": {
            "type": "string",
            "description": "ステータス",
            "enum": [
              "UNPUBLISHED",
              "PUBLISHED",
              "OUT_OF_PRINT"
            ],
            "example": "PUBLISHED"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "description": "出版日時:ISO8601",
            "example": "2025-01-23T01:23:45.000Z"
          },
          "authorId": {
            "type": "integer",
            "format": "int32",
            "description": "著者ID",
            "example": 1
          },
          "authorName": {
            "type": "string",
            "description": "著者名",
            "example": "Dustin Boswell"
          },
          "publisherId": {
            "type": "integer",
            "format": "int32",
            "description": "出版社ID",
            "example": 1
          },
          "publisherName": {
            "type": "string",
            "description": "出版社名",
            "example": "O\u0027Reilly"
          }
        },
        "required": [
          "id",
          "isbn",
          "status",
          "title"
        ]
      },
      "ResponsePagination": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int32",
            "description": "取得開始位置",
            "example": 0
          },
          "limit": {
            "type": "integer",
            "format": "int32",
            "description": "取得数",
            "example": 100
          },
          "total": {
            "type": "integer",
            "format": "int32",
            "description": "検索ヒット総数",
            "example": 1
          }
        },
        "required": [
          "limit",
          "offset",
          "total"
        ]
      },
      "SearchBooksResponse": {
        "type": "object",
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/ResponsePagination"
          },
          "hits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponseHit"
            }
          }
        }
      },
      "FindBookByIdResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "書籍ID",
            "example": 1
          },
          "isbn": {
            "type": "string",
            "description": "ISBN-13",
            "example": 9784873115658
          },
          "title": {
            "type": "string",
            "description": "タイトル",
            "example": "リーダブルコード"
          },
          "price": {
            "type": "integer",
            "format": "int32",
            "description": "価格",
            "example": 2640
          },
          "status": {
            "type": "string",
            "description": "ステータス",
            "enum": [
              "UNPUBLISHED",
              "PUBLISHED",
              "OUT_OF_PRINT"
            ],
            "example": "PUBLISHED"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "description": "出版日時:ISO8601",
            "example": "2025-01-23T01:23:45.000Z"
          },
          "authorId": {
            "type": "integer",
            "format": "int32",
            "description": "著者ID",
            "example": 1
          },
          "authorName": {
            "type": "string",
            "description": "著者名",
            "example": "Dustin Boswell"
          },
          "publisherId": {
            "type": "integer",
            "format": "int32",
            "description": "出版社ID",
            "example": 1
          },
          "publisherName": {
            "type": "string",
            "description": "出版社名",
            "example": "O\u0027Reilly"
          }
        },
        "required": [
          "id",
          "isbn",
          "status",
          "title"
        ]
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "ApiKeyAuth",
        "in": "header"
      }
    }
  }
}